split huge SelenideMethods to smaller tests
